+++ example/lib/pages/main_page.dart
@@ -1,9 +1,10 @@
+import 'package:collection/collection.dart';
 import 'package:example/example_routes.dart';
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/material.dart';
 
 import 'package:url_launcher/url_launcher.dart';
-import 'package:collection/collection.dart';
+
 import '../example_route.dart';
 import '../example_routes.dart' as example_routes;
 
@@ -23,8 +24,8 @@ class MainPage extends StatelessWidget {
             .where((FFRouteSettings element) => element.exts != null)
             .map<DemoRouteResult>((FFRouteSettings e) => DemoRouteResult(e))
             .toList()
-              ..sort((DemoRouteResult a, DemoRouteResult b) =>
-                  b.group.compareTo(a.group)),
+          ..sort((DemoRouteResult a, DemoRouteResult b) =>
+              b.group.compareTo(a.group)),
         (DemoRouteResult x) => x.group));
   }
   final Map<String, List<DemoRouteResult>> routesGroup =
@@ -51,7 +52,8 @@ class MainPage extends StatelessWidget {
                 ),
               ),
               onPressed: () {
-                launch('https://github.com/fluttercandies/extended_tabs');
+                launchUrl(Uri.parse(
+                    'https://github.com/fluttercandies/extended_tabs'));
               },
             ),
           ),
@@ -62,7 +64,7 @@ class MainPage extends StatelessWidget {
               child:
                   Image.network('https://pub.idqqimg.com/wpa/images/group.png'),
               onPressed: () {
-                launch('https://jq.qq.com/?_wv=1027&k=5bcc0gy');
+                launchUrl(Uri.parse('https://jq.qq.com/?_wv=1027&k=5bcc0gy'));
               },
             ),
           )