+++ lib/extended_tabs.dart
@@ -1,7 +1,9 @@
 library extended_tabs;
 
+export 'package:sync_scroll_library/sync_scroll_library.dart';
 export 'src/carousel_indicator.dart';
 export 'src/page_view.dart';
+export 'src/scrollable.dart';
 export 'src/tab_bar.dart';
 export 'src/tab_indicator.dart';
 export 'src/tab_view.dart';